Added support for packed sequences with flex attention
Sequence-packing allows one to pack multiple examples into a single
sequence. The attention mask is then used to prevent the model from
attending to tokens across example boundaries. This make much more
efficient use of resources, as you don't have to pad sequences (or pad
them less).

This also solves the problem of having to compromise on the
sequence length when there is high variability in the length of the
examples. For example, if the dataset has a few very long examples,
mixed with many short, examples, you no longer have to error towards the
median length to optimize compute. Just use a small batch size (possibly
1) with very long sequences.

This is made practical with the help of flex-attention, which allows for
sparse attention masking and computation.

- Cleaned up packed sequence support in block-tokenizer. It already had
  this feature, but it had suffered bit-rot over the last couple of
years.
- Added packed position-id encoding to data collator. When "packed" flag
  is set, it will identify EOS tokens and generate zero-indexed position
ids for each sub-sequence. This is the format expected by the HF masking
interface for generating masks for packed sequences.
- Fixed various issues related to packed sequence masking in
  causal_lm.py and the attention implementation.
- As the relative-position-attention module now can function without
  relative position, there's was not much utility in keeping the older
attention implementation around. The implementations have been merged
and the "causal-model" has been updated to support the various attention
implementations now. One less attention implementation to maintain!
- Added packed-sequence configuration to smollm-corpus/fineweb-edu for
  testing.
- Added packed-sequence configuration to Samantha dataset.
- Updated Samantha finetune project to include new packed sequence
  configurations.
- Added common trainer script option to select attention implementation.
  Ideally, packed sequences should use flex_attention, although SDPA and
eager should also work -- if you have enough memory.
- Fixed various issues surrounding introduction of new attention setting
  in configurations.
- Modified the base custom language model definition to integrate
  improved method of model construction from base model project. This
mostly needed to provide a uniform interface for setting attention type,
when constructing models.
- Cleaned up model definitions
- Added new "pretraining" example directory, with basic example using
  the smollm dataset and sequence packing. This is still a
work-in-progress.
- Added support to "latent.py" for deferring the construction of
  arguments, via "latent_args" argument. This allows a dictionary of
arguments to be passed in, at the time of materialization, and merges
them with the static arguments. This was needed to pass a dictionary of
model construction arguments (with attn_implementation) as an argument
to a sub-project.
- Modified a couple of the tokenizer configs to support a longer
  sequence lengths -- needed to properly test the new longer sequences
made practical with sequence packing.

TODO:
- Deepone model still is not fully integrated with the new attention
  system. It only support SDPA. This will require updating the ALiBi
attention implementation.
- Enable and test flash-attention-2 implementation.
- Update README's and documentation.
- Add more efficient packing to block_tokenizer implementation. It uses
  "greedy" packing, which is not ideal.
- Continue work on new pre-training examples.