Refactored attention mask from causal_lm
As we would like to support custom attention masks, the causal_mask code
from causal_lm.py has been factored out into a separate file. This will
make is much cleaner to add things like sliding window attention and
others in the very near future.

While working on this, the base template model was updated to pass the
model config directly to causal_lm, as this is much cleaner than setting
it as an attribute, externally.