Refactored model converter
[problem]
The model conversion tool was created before we had a model project
type, where the model definition was just a template inlined into a
training project. It had been hacked to resuse the same template,
outside of the context of a project.

Now that we have model projects, it should use the common definition for
constructing model instances.

[solution]
Refactored script to use Llama project for model construction.

Updated Llama model project to accept dynamic args -- which also allows
the converter to pass in the translated model parameters.

This is really the first phase in refactoring this code. Next, we will
want to abstract how to translate the model configuration and FQNs. In
theory, we should be able to add custom code in the model projects for
each model type.