Workaround for Qwen3's lack of BOS token
Qwen3 does not have a BOS token defined and we use this as a document
delimiter when packing multiple documents into a sequence with the
BlockTokenizer. The result is that Qwen3 does not work with packed
documents.

I fixed a couple of bugs surrounding not checking if bos_token is None.
This does not fix it, but at least does not crash now.

As a workaround, the converter has been updated to let you add (or
override) the special tokens: bos_token, eos_token, pad_token, and
unk_token and permits configuring how to initialize them, when creating
new tokens.

This allows adding a BOS token. This is just a temporary solution for
testing. I'll fix the actual issue in another commit, which is how be
communicate document boundaries from the tokenizer to the data-collator.

Having the ability to set these special tokens explicilty is also a
useful feature, even outside of the present context.