Hack to deal with token renaming
A common use-case is to change the EOS token to match the the ChatML
<|im_end|> token. The HF tokenizer API does not have a clean way of
doing this.

I put Claude to the task, who came up with a hack to just copy the old
EOS embedding to a newly added embedding. It's less that ideal, but
unblocks my testing workflow.

TODO: Revisit and do this right.