Added Jinja2 filter for Python to YAML
When rendering Python variables as YAML, via a Jinja2 template, simple
uses-cases work. For example, a str, bool, int, or even a list of int
will produce a result which can be parsed by YAML.

The problem is that there are cases which don't translate properly. For
example, None is not valid syntax and should be translated to 'null' or
~. Similarly, there not all Python float syntax will be recognized as a
float by YAML.

To address this, a Jinja2 filter has been added which will convert
Python variables to YAML compatible syntax.

The filter also accepts a default argument, as to allow it to act as a
replacement for the Jinja2 "default" filter, but producing valid YAML
syntax for the edge cases, like for None.

The syntax documentation has been updated to cover this new
functionality, with examples.