Harmonize Forgather config names with HF names
While we mostly use the HF conventions for configuration variable names,
there are still a handful of exceptions. This complicates converting
configurations between HF and FG format.

This is the first step in harmonizing the names, where the dynamic Llama
model definition has been updated to use the same names as HF configs.

We have also replaced the "default" filter instances with "toyaml,"
which automatically handles edge-cases when converting Python variables
to YAML syntax. This makes it easier to inject complex variables, like
"rope_scaling" into the template and have it render YAML compatible
syntax.