Fixed PP and fused-cross-entropy
Pipeline parallel was broken on account of not having updated the
refactored model to support the janky API for returning hidden-states,
rather than logits.

A more well defined API has been defined for getting hidden-states back from
the model, when using fused-cross-entropy. The trainers has been updated
to use it and this has fixed tiny_experiments/pipeline_parallel.

This intersects with the fused-corss-entropy, where it was discovered
that the CCE implementation does not work, unless the dtype is bfloat16
or float16 -- and the PyPi version does not support bias.

These issues have been resolved in the auto-detection logic and it has
been confirmed that this works with both the base trainer as well as PP.

Even still, the fallback involves torch.compiple(), which adds extra
time to training startup, which is annoying for quick tests. This
fused-cross-entropy functionality is also somewhat fragile.

To address this, the default is not to used fused-cross-entropy, but it
has been enabled by default for the "finetune" project definition, which
needs it for many use-cases.