Implemented vLLM Tensor Parallel support
The previous TP configuration had a few issues. It worked with the HF
implementation, but vLLM makes different assumptions about the model --
it assumes that the inner transformer module is named "model," where is
should be checking the base_model_prefix.

The solution is to just make the prefix explicit in the TP
configurations.

I also updated the attention layer to make the head dimension a
wild-card dimension when reshaping, as the number of heads now depends
upon the TP degree.