Don't use nn.Buffer for RoPE embeddings
A great many things don't handle tied buffers correctly. Examples:
- vLLM needs a hack to prevent loading them
- HF "device_map=auto"

The least complex way of handling this is to simply not make
RealRotaryPE a sub-class of nn.Module. To handle the case of
construction of the meta-device, we detect if the device has been moved
to a real device and construct and cache them.

To handle how "device_map='auto'" works, when the input device changes,
we move the cached buffer to the new device. This is not ideal, as it
has additional communication overhead. It may be worth caching a copy of
each device, but that's a project for another day and the present
abstraction should work.

Issues:
- This will almost certanly break PP auto-splitting
- This may cause issues where the model is traced on the meta device and
  run on the real device. As conditional code paths don't work, there
will be no way to construct it on the real device.

So far, it at least works withL vLLM, inference with
"device_map='auto'", normal training, and PP training; this may be
sufficient to justify the breakage.