Added support for vLLM pipeline parallel inference
The issues which needed to be overcome:
- vLLM makes a great many undocumented assumptions about the structure
of HF models. Figuring out what those assumptions are is a slow and
painful process of trial and error.
- vLLM assumes that all of the modules in the pp_plan belong to the
outer-most nn.Module. Our outer-module is just a HF compatibility shim,
over a specific model implementation -- CausalLM, for example. CausalLM
does not even directly own the layers, which are manaaged by the
layer-stack abstraction. You can't just give vLLM FQNs to the target
modules, as it is not calling ".get_module()," but using "getattr()" on
the outer model.
- vLLM assumes that exactly one of the modules named in pp_plan is an
  instance of nn.ModuleList; it assumes that this is where the model's
layers are stored. This assumption is broken, as we use nn.ModuleDict().
This is needed for working with Torch's Pipeline Parallel
implementation. By using a ModuleDict, rather than a ModuleList, you can
delete individual modules without any impact on the model's FQNs.
- vLLM does not delete unused modules, it replaces them with an instance
  of PPMissingLayer.

We addressed these issues by creating a virtual module interface, where
we intercept __getattr__() and __setattr__() to forward these to the correct
modules. In the case of the ModuleList issue, a proxy ModuleList is
created, which intercepts calls to __setitem__() by updating both the
proxy list and the underlying ModuleDict.

Testing:

- vLLMs PP appears to work with 2 GPUs. When tested with 3 and 4, it
appears to fail on account of attempting to get the input embeddings
_after_ it has replaced it with a PPMissingLayer. It seems likely that
this is a bug in vLLM, rather than anything we are doing wrong.

- TP still works. Including with 4 GPUs.
- Our inference server works
- Tested Forgather PP training
- Tested Forgather Trainer