Copy model generation configs on conversion
The model converted has been updated to copy the model's generation
config, when converted.

The inference client and server have been updated to respect the values
from the generation config, unless explicitly overridden.

The server now emits a warning if both temperature and top_p are both
set to != 1, as doing so tends to lead to excessive repetition.

As to reduce the log noise, printing the model has been moved from
"info" to "debug."