Trainer logging without TTY
[problem]
When the Trainer callback for basic logging, ProgressCallback, logs
messages, it logs them via TQDM. While this looks nice when running in
a TTY, it is "less-than-optimal" when logging to a file -- it looks
terrible and makes parsing far more difficult than need-be.

[solution]
If logging to a TTY, use TQDM, otherwise, just use Python's logger
library.

Some of the logging formatting code may be needed by other custom
loggers, so it has been moved into a new "logging.py" module.