Disable template cache in interactive mode
[problem]
The FG interactive shell caches the list of available configurations the
first time it is needed, then uses the cache for all subsequent
requests.

The result is that the shell must be restarted if the list of templates
changes.

[solution]
Disable the template cache and always get the current templates. It's
not terribly expensive -- at least not expensive enough to justify
the complexity of keeping track of modification times -- the alternative
solution.