Enable KV cache for ALiBI attention
The present ALiBI attention implementation does not support the KV
cache, which can lead to unexpected results when testing with the
inference server. It's possible to disable the cache to work around
this, but this will not be obvious to most people.

I have added support for the KV cache to the ALiBI attention
implementation.